+++ apps/ekstra-omsorgsdager-andre-forelder-ikke-tilsyn/src/app/søknad/steps/oppsummering/oppsummeringMessages.ts
@@ -4,19 +4,22 @@ const nb = {
         'Les gjennom oppsummeringen og sjekk at alt er riktig før du sender inn søknaden. Hvis du vil gjøre endringer, kan du gå tilbake.',
 
     'step.oppsummering.søker.header': 'Om deg som søker',
-    'step.oppsummering.søker.fnr': 'Fødselsnummer: {fødselsnummer}',
+    'step.oppsummering.søker.navn': 'Navn',
+    'step.oppsummering.søker.fnr': 'Fødselsnummer',
 
-    'step.oppsummering.omBarna.header': 'Dine barn',
-    'step.oppsummering.omBarna.barn': 'Barn',
+    'step.oppsummering.omBarna.header': 'Om barn',
+    'step.oppsummering.omBarna.barn': 'Dine barn',
     'step.oppsummering.omBarna.listItem': ' (fnr. {identitetsnummer})',
 
     'step.oppsummering.annenForelder.header': 'Om den andre forelderen',
-    'step.oppsummering.annenForelder.fnr': 'Fødselsnummer: {fødselsnummer}',
+    'step.oppsummering.annenForelder.navn': 'Navn',
+    'step.oppsummering.annenForelder.fnr': 'Fødselsnummer',
 
     'step.oppsummering.annenForelderensSituasjon.header': 'Den andre forelderens situasjon',
     'step.oppsummering.annenForelderensSituasjon.erVarighetMerEnn6Maneder':
         'Har du bekreftet at den andre forelderen ikke kan ha tilsyn med barn i en periode på minst 6 måneder?',
-    'step.oppsummering.annenForelderensSituasjon.tittel': 'Grunn',
+    'step.oppsummering.annenForelderensSituasjon.tittel':
+        'Grunn til at den andre forelderen ikke kan ha tilsyn med barn:',
     'step.oppsummering.annenForelderensSituasjon.beskrivelse': 'Beskrivelse av situasjonen:',
     'step.oppsummering.annenForelderensSituasjon.periode.header':
         'Periode når den andre forelderen ikke kan ha tilsyn med barnet/barna:',