+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/søknad/steps/oppsummering/components/DineBarnOppsummering.tsx
@@ -1,55 +1,100 @@
 import { BarnType } from '@navikt/sif-common-forms-ds';
-import { JaNeiSvar, SummaryBlock, SummaryList, SummarySection } from '@navikt/sif-common-ui';
+import { JaNeiSvar, SummaryList } from '@navikt/sif-common-ui';
 import { ISODateToDate, dateFormatter } from '@navikt/sif-common-utils';
-import { useAppIntl } from '../../../../i18n';
+import { AppText, useAppIntl } from '../../../../i18n';
 import { RegistrertBarn } from '../../../../types/RegistrertBarn';
-import { ApiBarn, RegistrertBarnTypeApi } from '../../../../types/søknadApiData/SøknadApiData';
+import { ApiBarn } from '../../../../types/søknadApiData/SøknadApiData';
 import { mapRegistrertBarnToApiBarn } from '../../../../utils/søknadsdataToApiData/getDineBarnApiDataFromSøknadsdata';
+import { FormSummary } from '@navikt/ds-react';
+import { EditStepLink } from '@navikt/sif-common-soknad-ds';
 
 interface Props {
     barn: ApiBarn[];
     registrerteBarn: RegistrertBarn[];
     harDeltBosted?: boolean;
+    onEdit?: () => void;
 }
 
-const DineBarnOppsummering = ({ barn, registrerteBarn, harDeltBosted }: Props) => {
+const DineBarnOppsummering = ({ barn, registrerteBarn, harDeltBosted, onEdit }: Props) => {
     const { text } = useAppIntl();
     const registrerteBarnSomIkkeSkalSendesInnMenVises = registrerteBarn.map(mapRegistrertBarnToApiBarn);
     return (
-        <SummarySection header={text('step.oppsummering.dineBarn.tittel')}>
-            <SummaryList
-                items={[...registrerteBarnSomIkkeSkalSendesInnMenVises, ...barn]}
-                itemRenderer={(barn: ApiBarn) => {
-                    return (
-                        <>
-                            <div>{barn.navn}</div>
-                            <div>
-                                {text('step.oppsummering.dineBarn.født', {
-                                    dato: dateFormatter.compact(ISODateToDate(barn.fødselsdato)),
-                                })}
-                            </div>
-                            {barn.type !== RegistrertBarnTypeApi.fraOppslag ? (
-                                <>
-                                    <div>
-                                        {text('step.oppsummering.dineBarn.id', {
-                                            identitetsnummer: barn.identitetsnummer,
-                                        })}
-                                    </div>
-                                    {barn.type === BarnType.fosterbarn && (
-                                        <div>{text('step.oppsummering.dineBarn.fosterbarn')}</div>
-                                    )}
-                                </>
-                            ) : undefined}
-                        </>
-                    );
-                }}
-            />
-            {harDeltBosted !== undefined && (
-                <SummaryBlock header={text('step.dineBarn.harDeltBosted.spm')}>
-                    <JaNeiSvar harSvartJa={harDeltBosted} />
-                </SummaryBlock>
-            )}
-        </SummarySection>
+        <FormSummary>
+            <FormSummary.Header>
+                <FormSummary.Heading level="2">
+                    <AppText id="step.oppsummering.dineBarn.tittel" />
+                </FormSummary.Heading>
+                {onEdit && <EditStepLink onEdit={onEdit} />}
+            </FormSummary.Header>
+            <FormSummary.Answers>
+                <FormSummary.Answer>
+                    <FormSummary.Label>
+                        <AppText id="step.oppsummering.dineBarn.registrerteBarn" />
+                    </FormSummary.Label>
+                    <FormSummary.Value>
+                        <SummaryList
+                            useAkselList={true}
+                            items={registrerteBarnSomIkkeSkalSendesInnMenVises}
+                            itemRenderer={(barn: ApiBarn) => {
+                                return (
+                                    <>
+                                        <div>{barn.navn}</div>
+                                        <div>
+                                            {text('step.oppsummering.dineBarn.født', {
+                                                dato: dateFormatter.compact(ISODateToDate(barn.fødselsdato)),
+                                            })}
+                                        </div>
+                                    </>
+                                );
+                            }}
+                        />
+                    </FormSummary.Value>
+                </FormSummary.Answer>
+                <FormSummary.Answer>
+                    <FormSummary.Label>
+                        <AppText id="step.oppsummering.dineBarn.andreBarn" />
+                    </FormSummary.Label>
+                    <FormSummary.Value>
+                        <SummaryList
+                            useAkselList={true}
+                            items={barn}
+                            itemRenderer={(barn: ApiBarn) => {
+                                return (
+                                    <>
+                                        <div>{barn.navn}</div>
+                                        <div>
+                                            {text('step.oppsummering.dineBarn.født', {
+                                                dato: dateFormatter.compact(ISODateToDate(barn.fødselsdato)),
+                                            })}
+                                        </div>
+
+                                        <div>
+                                            {text('step.oppsummering.dineBarn.id', {
+                                                identitetsnummer: barn.identitetsnummer,
+                                            })}
+                                        </div>
+                                        {barn.type === BarnType.fosterbarn && (
+                                            <div>{text('step.oppsummering.dineBarn.fosterbarn')}</div>
+                                        )}
+                                    </>
+                                );
+                            }}
+                        />
+                    </FormSummary.Value>
+                </FormSummary.Answer>
+
+                {harDeltBosted !== undefined && (
+                    <FormSummary.Answer>
+                        <FormSummary.Label>
+                            <AppText id="step.dineBarn.harDeltBosted.spm" />
+                        </FormSummary.Label>
+                        <FormSummary.Value>
+                            <JaNeiSvar harSvartJa={harDeltBosted} />
+                        </FormSummary.Value>
+                    </FormSummary.Answer>
+                )}
+            </FormSummary.Answers>
+        </FormSummary>
     );
 };