+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/søknad/steps/oppsummering/components/LegeerklæringOppsummering.tsx
@@ -1,34 +1,44 @@
-import React from 'react';
-import Block from '@navikt/sif-common-core-ds/src/atoms/block/Block';
+import { FormSummary } from '@navikt/ds-react';
 import AttachmentList from '@navikt/sif-common-core-ds/src/components/attachment-list/AttachmentList';
-import { SummarySection } from '@navikt/sif-common-ui';
-import { AppText, useAppIntl } from '../../../../i18n';
+import { EditStepLink } from '@navikt/sif-common-soknad-ds';
+import { AppText } from '../../../../i18n';
 import { LegeerklæringSøknadsdata } from '../../../../types/søknadsdata/LegeerklæringSøknadsdata';
 import { getAttachmentURLBackend } from '../../../../utils/attachmentUtilsAuthToken';
 
 interface Props {
     vedlegg: string[];
     legeerklæringSøknadsdata?: LegeerklæringSøknadsdata;
+    onEdit?: () => void;
 }
 
-const LegeerklæringOppsummering: React.FC<Props> = ({ vedlegg, legeerklæringSøknadsdata }) => {
-    const { text } = useAppIntl();
+const LegeerklæringOppsummering = ({ vedlegg, legeerklæringSøknadsdata, onEdit }: Props) => {
     const legeerklæringer = legeerklæringSøknadsdata
         ? legeerklæringSøknadsdata.vedlegg.filter((v) => v.url && vedlegg.includes(getAttachmentURLBackend(v.url)))
         : [];
 
     return (
-        <SummarySection header={text('step.oppsummering.legeerklæring.header')}>
-            <Block>
-                {legeerklæringSøknadsdata?.vedlegg.length === 0 ? (
-                    <AppText id="step.oppsummering.legeerklæring.ingenVedlegg" />
-                ) : (
-                    <div data-testid="legeerklæring-liste">
-                        <AttachmentList attachments={legeerklæringer} />
-                    </div>
-                )}
-            </Block>
-        </SummarySection>
+        <FormSummary>
+            <FormSummary.Header>
+                <FormSummary.Heading level="2">
+                    <AppText id="step.oppsummering.legeerklæring.header" />
+                </FormSummary.Heading>
+                {onEdit && <EditStepLink onEdit={onEdit} />}
+            </FormSummary.Header>
+            <FormSummary.Answers>
+                <FormSummary.Answer>
+                    <FormSummary.Label>
+                        <AppText id="step.oppsummering.legeerklæring.liste.header" />
+                    </FormSummary.Label>
+                    <FormSummary.Value>
+                        {legeerklæringSøknadsdata?.vedlegg.length === 0 ? (
+                            <AppText id="step.oppsummering.legeerklæring.ingenVedlegg" />
+                        ) : (
+                            <AttachmentList attachments={legeerklæringer} />
+                        )}
+                    </FormSummary.Value>
+                </FormSummary.Answer>
+            </FormSummary.Answers>
+        </FormSummary>
     );
 };