+++ apps/pleiepenger-i-livets-sluttfase-soknad/CHANGELOG.md
@@ -1,5 +1,35 @@
 # @navikt/pleiepenger-i-livets-sluttfase-soknad
 
+## 2.9.12
+
+### Patch Changes
+
+-   Bruke Aksel mal for oppsummeringssiden
+-   Updated dependencies
+    -   @navikt/sif-common-forms-ds@9.1.0
+
+## 2.9.11
+
+### Patch Changes
+
+-   Updated dependencies [a8ae0fe]
+    -   @navikt/sif-common-soknad-ds@12.0.10
+
+## 2.9.10
+
+### Patch Changes
+
+-   008a462: Pakkeoppdateringer - minor/patch
+-   Updated dependencies [008a462]
+    -   @navikt/sif-common-amplitude@2.25.15
+    -   @navikt/sif-common-formik-ds@1.11.6
+    -   @navikt/sif-common-soknad-ds@12.0.9
+    -   @navikt/sif-common-forms-ds@9.0.11
+    -   @navikt/sif-common-core-ds@9.17.24
+    -   @navikt/sif-common-sentry@0.26.17
+    -   @navikt/sif-common-utils@3.46.12
+    -   @navikt/sif-common-ui@0.7.29
+
 ## 2.9.9
 
 ### Patch Changes