+++ packages/sif-common-core-ds/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @navikt/sif-common-core-ds
 
+## 9.17.24
+
+### Patch Changes
+
+-   008a462: Pakkeoppdateringer - minor/patch
+-   Updated dependencies [008a462]
+    -   @navikt/sif-common-utils@3.46.12
+
 ## 9.17.23
 
 ### Patch Changes