+++ packages/sif-common-ui/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @navikt/sif-common-ui
 
+## 0.7.29
+
+### Patch Changes
+
+-   008a462: Pakkeoppdateringer - minor/patch
+-   Updated dependencies [008a462]
+    -   @navikt/sif-common-formik-ds@1.11.6
+    -   @navikt/sif-common-core-ds@9.17.24
+    -   @navikt/sif-common-utils@3.46.12
+
 ## 0.7.28
 
 ### Patch Changes