+++ packages/sif-common-utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/sif-common-utils
 
+## 3.46.12
+
+### Patch Changes
+
+-   008a462: Pakkeoppdateringer - minor/patch
+
 ## 3.46.11
 
 ### Patch Changes